chore: move asyncRequire to modules
