+++ packages/module-federation-metro/src/plugin.ts
@@ -382,7 +382,7 @@ function withModuleFederation(
   const remoteEntryPath = path.resolve(mfMetroPath, remoteEntryFilename);
   const remoteHMRSetupPath = path.resolve(mfMetroPath, 'remote-hmr.js');
 
-  const asyncRequirePath = path.resolve(__dirname, '../async-require.js');
+  const asyncRequirePath = path.resolve(__dirname, './modules/asyncRequire.js');
 
   const babelTransformerPath = createBabelTransformer({
     proxiedBabelTrasnsformerPath: config.transformer.babelTransformerPath,
